Add 'hdat' console entry point in setup.py
main.py is now the global entry point. The hdat command can be tested with "pip install -e .  && hdat", still possible to run with ./hdat/main.py
