+++ hdat/main.py
@@ -1,113 +1,41 @@
-import argparse
-import traceback
+#!/usr/bin/env python3
+import sys
+import os
 
-from .resultspec import resolve_resultspecs, print_resultspec
-from .casespec import resolve_casespecs, select_suite
-from .runner import run_cases
-from .util import AbortError
-
-
-def parse_arguments(arguments):
-    parser = argparse.ArgumentParser(prog='hdatt')
-    subparsers = parser.add_subparsers(dest='command', metavar='<command>')
-
-    run_help = 'run cases, store results in archive, compare against goldens'
-    run_parser = subparsers.add_parser('run', help=run_help)
-    run_parser.add_argument('casespecs', nargs='*', default=[''], metavar='<case>')
-    run_parser.add_argument('--collect-only', default=False, action='store_true')
-
-    show_help = 'visualize a single result'
-    show_parser = subparsers.add_parser('show', help=show_help)
-    show_result_help = 'result specifier to show'
-    show_parser.add_argument('resultspec', nargs="?", default='', metavar='<result>', help=show_result_help)
-
-    runshow_help = 'run then visualize a single result'
-    runshow_parser = subparsers.add_parser('runshow', help=runshow_help)
-    runshow_parser.add_argument('casespecs', nargs=1, default='', metavar='<result>')
-
-    diff_help = 'compare two results'
-    diff_parser = subparsers.add_parser('diff', help=diff_help)
-    diff_golden_help = 'result to compare to (defaults to current golden for the case)'
-    diff_parser.add_argument('goldenspec', nargs='?', default=None, metavar='<golden>', help=diff_golden_help)
-    diff_result_help = 'result being compared'
-    diff_parser.add_argument('resultspec', nargs=1, metavar='<result>', help=diff_result_help)
-
-    verify_help = 'move result metrics from archive to the golden store'
-    verify_parser = subparsers.add_parser('verify', help=verify_help)
-    verify_result_help = 'results to be stripped and moved into the golden store'
-    verify_parser.add_argument('resultspec', nargs='?', default='', metavar='<result>', help=verify_result_help)
-
-    return parser.parse_args(arguments)
+from hdat.hdat_cli import hdat_cli
+from hdat.suite import collect_suites
+from hdat.source_control import git_info_from_directory
+from hdat.util import repository_root, print_error, AbortError
+from hdat.store import Archive, GoldenStore
 
 
-def _format_cases_status(cases_status):
-    return 'PASS: {}, FAIL: {}, UNKNOWN: {}, ERROR: {}'.format(
-        cases_status['pass'],
-        cases_status['fail'],
-        cases_status['unknown'],
-        cases_status['error'],
-    )
+def main():
+    cwd = os.getcwd()
 
-
-def main(arguments, suites, golden_store, archive, git_info):
-    args = parse_arguments(arguments)
-
-    if args.command is None:
-        parse_arguments(['-h'])
-
-    if args.command == 'run' and args.collect_only:
-        cases = resolve_casespecs(suites, args.casespecs)
-        print("\n".join(['{}/{}'.format(s, c) for s, c in cases]))
-    elif args.command == 'run':
-        cases = resolve_casespecs(suites, args.casespecs)
-        cases_status = run_cases(suites, golden_store, archive, git_info, cases)
-        if cases_status['pass'] < len(cases):
-            raise AbortError(_format_cases_status(cases_status))
-    elif args.command == 'show':
-        results = resolve_resultspecs(archive, args.resultspec)
-        for result in results:
-            show_result(suites, result)
-    elif args.command == 'runshow':
-        cases = resolve_casespecs(suites, args.casespecs)
-        cases_status = run_cases(suites, golden_store, archive, git_info, cases)
-        if cases_status['error'] > 0:
-            raise AbortError(_format_cases_status(cases_status))
-        results = resolve_resultspecs(archive, args.casespecs[0])
-        for result in results:
-            show_result(suites, result)
-    elif args.command == 'diff':
-        golden_results = resolve_resultspecs(archive, args.goldenspec)
-        results = resolve_resultspecs(archive, args.resultspec)
-        for golden_result, result in zip(golden_results, results):
-            diff_results(suites, golden_result, result)
-    elif args.command == 'verify':
-        results = resolve_resultspecs(archive, args.resultspec)
-        for result in results:
-            golden_store.insert(result)
-
-
-def show_result(suites, result):
-    suite = select_suite(suites, result['suite_id'])
     try:
-        suite.show(result)
-    except Exception as e:
-        traceback.print_exc()
-        msg = 'Error when attempting to show "{}": {}'
-        raise AbortError(msg.format(print_resultspec(result), e))
+        git_info = git_info_from_directory(cwd)
+        repo_directory = repository_root(cwd)
+
+        if 'HDATT_ARCHIVE' in os.environ:
+            archive_location = os.environ['HDATT_ARCHIVE']
+        else:
+            archive_location = os.path.join(repo_directory, '.hdattarchive')
+        archive = Archive(archive_location)
+
+        golden_store_location = os.path.join(repo_directory, 'golden_results')
+        golden_store = GoldenStore(golden_store_location)
+
+        suites = collect_suites(cwd)
+
+        hdat_cli(sys.argv[1:], suites, golden_store, archive, git_info)
+
+        sys.exit(0)
+
+    except AbortError as e:
+        print_error(e)
+
+        sys.exit(1)
 
 
-def diff_results(suites, golden_result, result):
-    suite_id = result['suite_id']
-    golden_suite_id = golden_result['result_id']
-
-    if golden_suite_id != suite_id:
-        msg = 'Can not diff results from different suites "{}" and "{}"'
-        raise AbortError(msg.format(golden_suite_id, suite_id))
-
-    suite = select_suite(suites, suite_id)
-    try:
-        suite.diff(golden_result, result)
-    except Exception as e:
-        traceback.print_exc()
-        msg = 'Error when attempting to show "{}": {}'
-        raise AbortError(msg.format(print_resultspec(result), e))
+if __name__ == '__main__':
+    main()
