--- hdat_main_script
@@ -1,37 +1,0 @@
-#!/usr/bin/env python
-import sys
-import os
-
-from hdatt.main import main
-from hdatt.suite import collect_suites
-from hdatt.source_control import git_info_from_directory
-from hdatt.util import repository_root, print_error, AbortError
-from hdatt.store import Archive, GoldenStore
-
-
-if __name__ == '__main__':
-    cwd = os.getcwd()
-
-    try:
-        git_info = git_info_from_directory(cwd)
-        repo_directory = repository_root(cwd)
-
-        if 'HDATT_ARCHIVE' in os.environ:
-            archive_location = os.environ['HDATT_ARCHIVE']
-        else:
-            archive_location = os.path.join(repo_directory, '.hdattarchive')
-        archive = Archive(archive_location)
-
-        golden_store_location = os.path.join(repo_directory, 'golden_results')
-        golden_store = GoldenStore(golden_store_location)
-
-        suites = collect_suites(cwd)
-
-        main(sys.argv[1:], suites, golden_store, archive, git_info)
-
-        sys.exit(0)
-
-    except AbortError as e:
-        print_error(e)
-
-        sys.exit(1)
